Reworked test app, so that each control has it's own demo page.
