+++ src/Callisto.TestApp/App.xaml.cs
@@ -48,7 +48,7 @@ namespace XamlControlsUITestApp
 
             // Create a Frame to act navigation context and navigate to the first page
             var rootFrame = new Frame();
-            rootFrame.Navigate(typeof(BlankPage));
+            rootFrame.Navigate(typeof(Callisto.TestApp.MainPage));
 
             // Place the frame in the current Window and ensure that it is active
             Window.Current.Content = rootFrame;
