+++ src/Callisto.TestApp/Callisto.TestApp.csproj
@@ -125,10 +125,28 @@
     <Compile Include="BlankPage.xaml.cs">
       <DependentUpon>BlankPage.xaml</DependentUpon>
     </Compile>
+    <Compile Include="MainPage.xaml.cs">
+      <DependentUpon>MainPage.xaml</DependentUpon>
+    </Compile>
     <Compile Include="ObjectTracker.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SampleInput.xaml.cs">
       <DependentUpon>SampleInput.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="SamplePages\FlyoutSample.xaml.cs">
+      <DependentUpon>FlyoutSample.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="SamplePages\LiveTileSample.xaml.cs">
+      <DependentUpon>LiveTileSample.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="SamplePages\SettingsSample.xaml.cs">
+      <DependentUpon>SettingsSample.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="SamplePages\TiltSample.xaml.cs">
+      <DependentUpon>TiltSample.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="StartPage.xaml.cs">
+      <DependentUpon>StartPage.xaml</DependentUpon>
     </Compile>
   </ItemGroup>
   <ItemGroup>
@@ -154,10 +172,40 @@
     <Content Include="Images\butterfly.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+    <Content Include="Images\Callisto.png">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Images\CallistoWhite.png">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
     <Content Include="Images\grasshopper.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+    <Page Include="MainPage.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
     <Page Include="SampleInput.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="SamplePages\FlyoutSample.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="SamplePages\LiveTileSample.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="SamplePages\SettingsSample.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="SamplePages\TiltSample.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="StartPage.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
