+++ src/Callisto.TestApp/SamplePages/FlyoutSample.xaml
@@ -1,0 +1,61 @@
+﻿<Page
+    x:Class="Callisto.TestApp.SamplePages.FlyoutSample"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Callisto.TestApp.SamplePages"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+	<Grid  x:Name="LayoutRoot" Background="{StaticResource ApplicationPageBackgroundBrush}">
+		<StackPanel Margin="100,100,0,0" Width="500" HorizontalAlignment="Left">
+			<TextBlock Text="Menu Position" />
+			<ComboBox x:Name="positioning" 
+					  Margin="22,0,22,0"
+					  Height="55" >
+				<ComboBoxItem>Left</ComboBoxItem>
+				<ComboBoxItem>Right</ComboBoxItem>
+				<ComboBoxItem>Top</ComboBoxItem>
+				<ComboBoxItem>Bottom</ComboBoxItem>
+			</ComboBox>
+			
+			<Button Margin="20,20,0,0" Content="Show Menu" 
+					Click="ShowFlyoutMenu2" x:Name="middlebutton" />
+
+			<TextBlock Text="App bar menu event log" Margin="0,20,0,0" />
+			<TextBox x:Name="LogOutput" AcceptsReturn="True" Margin="22,0" />
+		</StackPanel>
+	</Grid>
+	
+	<Page.BottomAppBar>
+		<AppBar Height="88" Background="{StaticResource AppBarTestBackground}">
+			<Grid>
+				<StackPanel HorizontalAlignment="Left" Orientation="Horizontal">
+					<Button Margin="20,10,0,5" Content="AppBar Left bottom" x:Name="leftbutton" Click="ShowFlyoutMenu" />
+					<Button Style="{StaticResource MailAppBarButtonStyle}" Click="ShowFlyoutMenu"/>
+					<Button Style="{StaticResource PinAppBarButtonStyle}" Click="ShowFlyoutMenu"/>
+					<Button Style="{StaticResource UnPinAppBarButtonStyle}" Click="ShowFlyoutMenu"/>
+				</StackPanel>
+				<StackPanel HorizontalAlignment="Right">
+					<Button Margin="0,0,20,0" Style="{StaticResource AddAppBarButtonStyle}" Click="ShowFlyoutMenu3" />
+				</StackPanel>
+			</Grid>
+		</AppBar>
+	</Page.BottomAppBar>
+	
+	<Page.TopAppBar>
+		<AppBar Height="88" Background="{StaticResource AppBarTestBackground}">
+			<Grid>
+				<StackPanel HorizontalAlignment="Left">
+					<Button Margin="20,0,0,0" Content="AppBar Left top" x:Name="leftbuttontop" Click="ShowFlyoutMenu" />
+				</StackPanel>
+				<StackPanel HorizontalAlignment="Right" Orientation="Horizontal">
+					<Button Style="{StaticResource MailAppBarButtonStyle}" Click="ShowFlyoutMenu"/>
+					<Button Style="{StaticResource PinAppBarButtonStyle}" Click="ShowFlyoutMenu"/>
+					<Button Style="{StaticResource UnPinAppBarButtonStyle}" Click="ShowFlyoutMenu"/>
+					<Button Margin="0,0,20,0" Content="AppBar Right top" Click="ShowFlyoutMenu" />
+				</StackPanel>
+			</Grid>
+		</AppBar>
+	</Page.TopAppBar>
+</Page>
