+++ src/Callisto.TestApp/SamplePages/LiveTileSample.xaml
@@ -1,0 +1,100 @@
+﻿<Page
+    x:Class="Callisto.TestApp.SamplePages.LiveTileSample"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Callisto.TestApp.SamplePages"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+     xmlns:callisto="using:Callisto.Controls"
+	mc:Ignorable="d">
+
+	<Grid Background="{StaticResource ApplicationPageBackgroundBrush}">
+
+		<Grid.ColumnDefinitions>
+			<ColumnDefinition Width="245" />
+			<ColumnDefinition Width="155" />
+		</Grid.ColumnDefinitions>
+		<Grid.RowDefinitions>
+			<RowDefinition Height="155" />
+			<RowDefinition Height="155" />
+			<RowDefinition Height="155" />
+		</Grid.RowDefinitions>
+
+		<callisto:LiveTile x:Name="liveTile"
+						   ItemsSource="{Binding}"
+						   BorderBrush="White" BorderThickness="1"
+						   Margin="5" Background="LightBlue">
+			<callisto:LiveTile.ItemTemplate>
+				<DataTemplate>
+					<Grid>
+						<Image Stretch="UniformToFill" >
+							<Image.Source>
+								<BitmapImage UriSource="{Binding ImageUri}" />
+							</Image.Source>
+						</Image>
+						<Border Background="#99000000" VerticalAlignment="Bottom">
+							<StackPanel Margin="5">
+								<TextBlock Foreground="White" Text="{Binding Name}" FontWeight="Bold" />
+								<TextBlock Foreground="White" Text="{Binding Description}"
+										   TextWrapping="Wrap"
+										   MaxHeight="30" TextTrimming="WordEllipsis"  />
+							</StackPanel>
+						</Border>
+					</Grid>
+				</DataTemplate>
+			</callisto:LiveTile.ItemTemplate>
+		</callisto:LiveTile>
+
+
+		<callisto:LiveTile x:Name="liveTile2"
+						   ItemsSource="{Binding}"
+						   Grid.ColumnSpan="2"
+						   Grid.Row="1"
+						   BorderBrush="White" BorderThickness="1"
+						   Margin="5" Background="#FF333333"
+						   Direction="Left">
+			<callisto:LiveTile.ItemTemplate>
+				<DataTemplate>
+					<Grid Margin="5">
+						<Grid.RowDefinitions>
+							<RowDefinition Height="Auto" />
+							<RowDefinition Height="*" />
+							<RowDefinition Height="Auto" />
+						</Grid.RowDefinitions>
+						
+						<TextBlock Text="{Binding Name}" 
+									Foreground="White" FontSize="24"
+									FontFamily="Segoe UI Light" />
+						
+						<TextBlock Foreground="White" Text="{Binding Description}"
+									TextWrapping="Wrap"
+									TextTrimming="WordEllipsis" 
+									FontSize="14"
+									Grid.Row="1" />
+
+						<HyperlinkButton Content="read more..."
+									   Grid.Row="2" 
+									   HorizontalAlignment="Right"
+									   Tag="{Binding ReadMoreUri}"
+									   Foreground="LightGray"
+									   Click="OnReadMoreLink_Click" />
+					</Grid>
+				</DataTemplate>
+			</callisto:LiveTile.ItemTemplate>
+		</callisto:LiveTile>
+
+		<callisto:LiveTile x:Name="liveTile3"
+					ItemsSource="{Binding}"
+					Grid.Column="2" Grid.Row="2" Margin="5">
+			<callisto:LiveTile.ItemTemplate>
+				<DataTemplate>
+					<Image Stretch="UniformToFill" >
+						<Image.Source>
+							<BitmapImage UriSource="{Binding ImageUri}" />
+						</Image.Source>
+					</Image>
+				</DataTemplate>
+			</callisto:LiveTile.ItemTemplate>
+		</callisto:LiveTile>
+	</Grid>
+</Page>
