+++ src/Callisto.TestApp/SamplePages/SettingsSample.xaml
@@ -1,0 +1,20 @@
+﻿<Page
+    x:Class="Callisto.TestApp.SamplePages.SettingsSample"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Callisto.TestApp.SamplePages"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Grid Background="{StaticResource ApplicationPageBackgroundBrush}">
+		<StackPanel Margin="100,100,0,0" HorizontalAlignment="Left">
+			<TextBlock Text="Settings Width" />
+			<ComboBox x:Name="settingswidth" Width="400" Height="55" HorizontalAlignment="Left">
+				<ComboBoxItem>Narrow</ComboBoxItem>
+				<ComboBoxItem>Wide</ComboBoxItem>
+			</ComboBox>
+			<Button Content="Show Settings" Click="ShowSettings" />
+		</StackPanel>
+	</Grid>
+</Page>
