+++ src/Callisto.TestApp/SamplePages/TiltSample.xaml
@@ -1,0 +1,94 @@
+﻿<Page
+    x:Class="Callisto.TestApp.SamplePages.TiltSample"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Callisto.TestApp.SamplePages"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	xmlns:callistoEffects="using:Callisto.Effects"
+	xmlns:sys="using:System"
+    mc:Ignorable="d"
+	>
+
+    <Grid Background="{StaticResource ApplicationPageBackgroundBrush}"
+		  Width="500" Height="400"
+		  HorizontalAlignment="Left"
+		  VerticalAlignment="Top"
+		  >
+		<Grid.Resources>
+			<Style TargetType="Border">
+				<Setter Property="Margin" Value="10" />
+				<Setter Property="BorderBrush" Value="White" />
+				<Setter Property="BorderThickness" Value="2" />
+			</Style>
+			<Style TargetType="TextBlock">
+				<Setter Property="HorizontalAlignment" Value="Center" />
+				<Setter Property="VerticalAlignment" Value="Center" />
+				<Setter Property="FontSize" Value="30" />
+			</Style>
+		</Grid.Resources>
+		<Grid.ColumnDefinitions>
+			<ColumnDefinition />
+			<ColumnDefinition />
+		</Grid.ColumnDefinitions>
+		<Grid.RowDefinitions>
+			<RowDefinition />
+			<RowDefinition />
+			<RowDefinition />
+		</Grid.RowDefinitions>
+
+		<Border Background="Red"
+				callistoEffects:Tilt.IsTiltEnabled="True">
+			<TextBlock Text="Tap me" />
+		</Border>
+
+		<Border Background="Green" Grid.Column="1"
+				callistoEffects:Tilt.IsTiltEnabled="True">
+			<TextBlock Text="Tap me" />
+		</Border>
+
+		<Grid  Grid.Column="0" Grid.Row="1" Margin="10"
+				callistoEffects:Tilt.IsTiltEnabled="True">
+			<Ellipse Fill="DarkViolet" Stroke="White" StrokeThickness="2" >
+			</Ellipse>
+			<TextBlock Text="Tap me" />
+		</Grid>
+
+		<Border Background="DarkCyan" Grid.Row="2" Grid.ColumnSpan="2"
+				callistoEffects:Tilt.IsTiltEnabled="True">
+			<TextBlock Text="Tap me" />
+		</Border>
+		
+		<!--<GridView IsItemClickEnabled="False" IsSwipeEnabled="False" IsTapEnabled="False">
+			<GridView.Items>
+				<Color>Red</Color>
+				<Color>Green</Color>
+				<Color>Blue</Color>
+				<Color>Orange</Color>
+				<Color>Magenta</Color>
+				<Color>Red</Color>
+				<Color>Green</Color>
+				<Color>Blue</Color>
+				<Color>Orange</Color>
+				<Color>Magenta</Color>
+				<Color>Red</Color>
+				<Color>Green</Color>
+				<Color>Blue</Color>
+				<Color>Orange</Color>
+				<Color>Magenta</Color>
+				<Color>Red</Color>
+			</GridView.Items>
+			<GridView.ItemTemplate>
+				<DataTemplate>
+					<Rectangle Width="240" Height="150"
+							   callistoEffects:Tilt.IsTiltEnabled="True">
+						<Rectangle.Fill>
+							<SolidColorBrush Color="{Binding}" />
+						</Rectangle.Fill>
+					</Rectangle>
+				</DataTemplate>
+			</GridView.ItemTemplate>
+		</GridView>-->
+		
+    </Grid>
+</Page>
