+++ src/Callisto.TestApp/StartPage.xaml
@@ -1,0 +1,15 @@
+﻿<Page
+    x:Class="Callisto.TestApp.StartPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:Callisto.TestApp"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Grid Background="{StaticResource ApplicationPageBackgroundBrush}">
+			<TextBlock Text="Select a sample on your right"
+					   VerticalAlignment="Center"
+					   HorizontalAlignment="Center" />
+    </Grid>
+</Page>
